preciowatt_v3_flask/requirements.txt
Flask>=3.0,<4.0
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
pyarrow>=14.0
pydataxm>=0.2
matplotlib>=3.7
boto3>=1.34
python-dotenv>=1.0
gunicorn>=21.2
